Create Graphical Display V1.py
Initial code commit.
There are a few functions that need to be written still:
    nearest
    near
    cost
    swap

As well as a data structure needs to be picked to correctly create the tree/graph object used within the bi_rrt_start_body function and the bi_rrt_star_extend function. This will likely have to be a custom class but could be done with an import and adding functionality to existing methods.